+++ __tests__/app-5-6-auth.js
@@ -13,7 +13,7 @@ describe('generator-laravel-5:app', () => {
         proxy: 'localhost',
         schema: 'http',
         preset: 'none',
-        enableAuth: false,
+        enableAuth: true,
         localGit: false
       })
       .on('end', done);
@@ -60,6 +60,20 @@ describe('generator-laravel-5:app', () => {
         '  }\n' +
         '});'
     );
+
+    done();
+  });
+
+  it('creates laravel auth layer files', done => {
+    assert.file([
+      'app/Http/Controllers/HomeController.php',
+      'resources/views/auth/passwords/email.blade.php',
+      'resources/views/auth/passwords/reset.blade.php',
+      'resources/views/auth/login.blade.php',
+      'resources/views/auth/register.blade.php',
+      'resources/views/home.blade.php',
+      'resources/views/layouts/app.blade.php'
+    ]);
 
     done();
   });