+++ __tests__/app-5-7-preset-vue.js
@@ -9,10 +9,10 @@ describe('generator-laravel-5:app', () => {
       .run(path.join(__dirname, '../generators/app'))
       .withPrompts({
         appname: 'testapp',
-        version: '5.6.*',
+        version: '5.7.*',
         proxy: 'localhost',
         schema: 'http',
-        preset: 'none',
+        preset: 'vue',
         enableAuth: false,
         localGit: false
       })
@@ -40,14 +40,27 @@ describe('generator-laravel-5:app', () => {
     done();
   });
 
+  it('packages.json contains npm packages', done => {
+    assert.fileContent('package.json', 'axios');
+    assert.fileContent('package.json', 'bootstrap');
+    assert.fileContent('package.json', 'bootstrap-sass');
+    assert.fileContent('package.json', 'browser-sync');
+    assert.fileContent('package.json', 'browser-sync-webpack-plugin');
+    assert.fileContent('package.json', 'cross-env');
+    assert.fileContent('package.json', 'jquery');
+    assert.fileContent('package.json', 'laravel-mix');
+    assert.fileContent('package.json', 'lodash');
+    assert.fileContent('package.json', 'popper.js');
+    assert.fileContent('package.json', 'vue');
+
+    done();
+  });
+
   it('webpack.mix.js is configured correctly', done => {
+    assert.fileContent('webpack.mix.js', "mix.js('resources/js/app.js', 'public/js');");
     assert.fileContent(
       'webpack.mix.js',
-      "mix.js('resources/assets/js/app.js', 'public/js');"
-    );
-    assert.fileContent(
-      'webpack.mix.js',
-      "mix.sass('resources/assets/sass/app.scss', 'public/css');"
+      "mix.sass('resources/sass/app.scss', 'public/css');"
     );
     assert.fileContent(
       'webpack.mix.js',